+++ src/errors.rs
@@ -6,6 +6,7 @@ use crate::{
     ckb::FundingError,
     fiber::{
         channel::{ChannelActorMessage, ProcessingChannelError},
+        graph::GraphError,
         types::Hash256,
         NetworkActorMessage,
     },
@@ -37,6 +38,8 @@ pub enum Error {
     FundingError(#[from] FundingError),
     #[error("InvalidParameter: {0}")]
     InvalidParameter(String),
+    #[error("Network Graph error: {0}")]
+    NetworkGraphError(#[from] GraphError),
 }
 
 pub type Result<T> = std::result::Result<T, Error>;
